chore: domain-infra 계층 분리. domain 계층은 비즈니스 모델과 로직만 관리.
