--- src/main/kotlin/scope/commerce/domain/product/infra/entity/Product.kt
@@ -1,13 +1,0 @@
-package scope.commerce.domain.product.infra.entity
-
-import jakarta.persistence.*
-
-@Entity
-class Product(
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    val id: Int = 0,
-    var name: String,
-    var price: Int,
-    var quantity: Int
-)
