+++ src/main/kotlin/scope/commerce/infra/bucket/entity/Bucket.kt
@@ -1,7 +1,7 @@
-package scope.commerce.domain.bucket.infra.entity
+package scope.commerce.infra.bucket.entity
 
 import jakarta.persistence.*
-import scope.commerce.domain.user.infra.entity.User
+import scope.commerce.infra.user.entity.User
 import java.time.LocalDateTime
 
 @Entity
