+++ src/main/kotlin/scope/commerce/infra/product/entity/Product.kt
@@ -1,12 +1,13 @@
-package scope.commerce.domain.user.infra.entity
+package scope.commerce.infra.product.entity
 
 import jakarta.persistence.*
 
 @Entity
-@Table(name = "users")
-class User(
+class Product(
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     val id: Int = 0,
-    var name: String
+    var name: String,
+    var price: Int,
+    var quantity: Int
 )
