+++ src/main/kotlin/scope/commerce/infra/product/entity/ProductStock.kt
@@ -1,17 +1,17 @@
-package scope.commerce.domain.coupon.infra.entity
+package scope.commerce.infra.product.entity
 
 import jakarta.persistence.*
 import java.time.LocalDateTime
 
 @Entity
-@Table(name = "coupon_stock")
-class CouponStock(
+@Table(name = "product_stock")
+class ProductStock(
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     val id: Long = 0,
     @OneToOne
-    @JoinColumn(name = "coupon_id", nullable = false, unique = true)
-    val coupon: Coupon,
-    val remainingQuantity: Int,
+    @JoinColumn(name = "product_id", nullable = false, unique = true)
+    val product: Product,
+    val quantity: Int,
     var updatedAt: LocalDateTime = LocalDateTime.now()
 )
