+++ src/main/kotlin/scope/commerce/infra/user/entity/User.kt
@@ -1,0 +1,12 @@
+package scope.commerce.infra.user.entity
+
+import jakarta.persistence.*
+
+@Entity
+@Table(name = "users")
+class User(
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    val id: Int = 0,
+    var name: String
+)
